streamprocessor/requirements.txt
kafka-python>=2.0.0
clickhouse-driver>=0.2.0
pandas>=1.5.0
pyarrow>=10.0.0
python-dotenv>=0.21.0
jsonschema>=4.0.0
boto3>=1.28.0
requests>=2.28.0
quixstreams
logfire